requirements.txt
python-telegram-bot==13.15
flask==2.2.3
gunicorn==20.1.0
python-dotenv==1.0.0
requests==2.28.2 